Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/websockets/interfaces/WebSocket/readyState/003-expected.txt

CONSOLE MESSAGE: line 12: WebSocket connection to 'ws://localhost:49001/' failed: WebSocket is closed before the connection is established.

FAIL WebSockets: delete readyState assert_equals: delete ws.readyState expected 2 but got 3
